不要なファイルを削除し、微修正

+++ app/Controller/RolesController.php
@@ -8,6 +8,7 @@ App::uses('AppController', 'Controller');
  */
 class RolesController extends AppController {
 
+	const ADMIN_ID = 1;
 /**
  * Components
  *
@@ -20,11 +21,23 @@ class RolesController extends AppController {
  *
  * @return void
  */
-	public function index() {
-		$this->Role->recursive = 0;
-		$this->set('roles', $this->Paginator->paginate());
+	public function index($id = null) {
+		if ($this->request->is('post')) {
+			if (is_null($id)) {
+				$this->add();
+				return;
+			}
+			$this->edit($id);
+			return;
+		} elseif ($this->request->is('delete')) {
+			$this->delete($id);
+			return;
+		} elseif ($this->request->is('get')) {
+			$this->search();
+			return;
+		}
+		$this->search();
 	}
-
 /**
  * view method
  *
@@ -34,7 +47,8 @@ class RolesController extends AppController {
  */
 	public function view($id = null) {
 		if (!$this->Role->exists($id)) {
-			throw new NotFoundException(__('Invalid role'));
+			$this->Flash->error("そのIDは存在しません。");
+			return $this->redirect(array('action' => 'index'));
 		}
 		$options = array('conditions' => array('Role.' . $this->Role->primaryKey => $id));
 		$this->set('role', $this->Role->find('first', $options));
@@ -55,7 +69,7 @@ class RolesController extends AppController {
 				$this->Flash->error(__('The role could not be saved. Please, try again.'));
 			}
 		}
-		$roles = $this->Role->Role->find('list');
+		$roles = $this->Role->find('list');
 		$this->set(compact('roles'));
 	}
 
@@ -68,8 +82,14 @@ class RolesController extends AppController {
  */
 	public function edit($id = null) {
 		if (!$this->Role->exists($id)) {
-			throw new NotFoundException(__('Invalid role'));
+			$this->Flash->error("そのIDは存在しません。");
+			return $this->redirect(array('action' => 'index'));
 		}
+		if ($id == self::ADMIN_ID) {
+			$this->Flash->error("そのIDは編集できません。");
+			return $this->redirect(array('action' => 'index'));
+		}
+
 		if ($this->request->is(array('post', 'put'))) {
 			if ($this->Role->save($this->request->data)) {
 				$this->Flash->success(__('The role has been saved.'));
@@ -81,7 +101,7 @@ class RolesController extends AppController {
 			$options = array('conditions' => array('Role.' . $this->Role->primaryKey => $id));
 			$this->request->data = $this->Role->find('first', $options);
 		}
-		$roles = $this->Role->Role->find('list');
+		$roles = $this->Role->find('list');
 		$this->set(compact('roles'));
 	}
 
@@ -94,14 +114,56 @@ class RolesController extends AppController {
  */
 	public function delete($id = null) {
 		if (!$this->Role->exists($id)) {
-			throw new NotFoundException(__('Invalid role'));
+			$this->Flash->error("そのIDは存在しません。");
+			return $this->redirect(array('action' => 'index'));
 		}
-		$this->request->allowMethod('post', 'delete');
+		if ($id == self::ADMIN_ID) {
+			$this->Flash->error("そのIDは削除できません。");
+			return $this->redirect(array('action' => 'index'));
+		}
+		$this->request->allowMethod('post', 'delete', 'get');
 		if ($this->Role->delete($id)) {
 			$this->Flash->success(__('The role has been deleted.'));
 		} else {
 			$this->Flash->error(__('The role could not be deleted. Please, try again.'));
 		}
 		return $this->redirect(array('action' => 'index'));
+	}
+
+		/**
+	 * Role一覧取得Ajax
+	 * @param int $per_page 1ページのデータ数
+	 * @param int $page 何ページ目か
+ 	 */
+	  public function search() {
+		$page = Hash::get($this->request->query, "page", self::DEFALUT_PAGE);
+		$per_page = Hash::get($this->request->query, "per_page", self::DEFALUT_PER_PAGE);
+		$roleId = Hash::get($this->request->query, "role_id");
+		$roleName = Hash::get($this->request->query, "role_name");
+		$adminFlag = Hash::get($this->request->query, "admin_flag");
+		$conditions = array(
+			'Role.id' => $roleId,
+			'Role.role_name' => $roleName,
+			'Role.admin_flag' => $adminFlag,
+		);
+
+		if (is_null($roleId) && is_null($roleName) && is_null($adminFlag)) {
+			$conditions = array();
+		}
+		$paginate =	array(
+			'page' => $page,
+			'fields' => array('id', 'role_name', 'admin_flag', 'created', 'modified',),
+			'limit' => $per_page,
+			'conditions' => array(
+				'or' => $conditions,
+			),
+			'order' => array(
+				'Role.id' => 'asc',
+			)
+		);
+		$this->Paginator->settings = $paginate;
+		$this->setJsonResponce($this->Paginator->paginate());
+		$this->render('index');
+
 	}
 }
